fix: validate workflow launch readiness
